requirements.txt
pandas
nltk
pyspellchecker
scikit-learn
flask
joblib